# apps/be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/catalog-nodes": {
      "post": {
        "operationId": "CatalogNodes_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatalogNodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNode"
                }
              }
            }
          }
        },
        "summary": "Create logical catalog node type (slug)",
        "tags": [
          "Catalog - nodes"
        ]
      },
      "get": {
        "operationId": "CatalogNodes_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Column to sort by (resource-specific whitelist)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ASC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text search (maps to searchable columns per endpoint)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCatalogNodes"
                }
              }
            }
          }
        },
        "summary": "List catalog nodes (paginated)",
        "tags": [
          "Catalog - nodes"
        ]
      }
    },
    "/catalog-nodes/{id}": {
      "get": {
        "operationId": "CatalogNodes_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNode"
                }
              }
            }
          }
        },
        "summary": "Get catalog node by id",
        "tags": [
          "Catalog - nodes"
        ]
      },
      "patch": {
        "operationId": "CatalogNodes_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatalogNodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNode"
                }
              }
            }
          }
        },
        "summary": "Update catalog node",
        "tags": [
          "Catalog - nodes"
        ]
      },
      "delete": {
        "description": "Hard delete; cascades to versions when no FKs block it.",
        "operationId": "CatalogNodes_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Delete catalog node type",
        "tags": [
          "Catalog - nodes"
        ]
      }
    },
    "/catalog-nodes/{catalogNodeId}/versions": {
      "get": {
        "operationId": "CatalogNodeVersionsList_findAll",
        "parameters": [
          {
            "name": "catalogNodeId",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Column to sort by (resource-specific whitelist)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ASC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text search (maps to searchable columns per endpoint)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "includeDeprecated",
            "required": false,
            "in": "query",
            "description": "When true, include deprecated versions (deprecatedAt set)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCatalogNodeVersions"
                }
              }
            }
          }
        },
        "summary": "List catalog versions for a node type (paginated)",
        "tags": [
          "Catalog - nodes"
        ]
      }
    },
...
    "/catalog-node-versions": {
      "post": {
        "operationId": "CatalogNodeVersions_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatalogNodeVersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeVersion"
                }
              }
            }
          }
        },
        "summary": "Create immutable catalog version for a node type",
        "tags": [
          "Catalog - node versions"
        ]
      }
    },
    "/catalog-node-versions/{id}": {
      "get": {
        "operationId": "CatalogNodeVersions_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeVersion"
                }
              }
            }
          }
        },
        "summary": "Get catalog node version by id",
        "tags": [
          "Catalog - node versions"
        ]
      },
      "patch": {
        "operationId": "CatalogNodeVersions_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatalogNodeVersionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeVersion"
                }
              }
            }
          }
        },
        "summary": "Update catalog node version (metadata / flags)",
        "tags": [
          "Catalog - node versions"
        ]
      },
      "delete": {
        "description": "Sets deprecatedAt and isActive=false; returns the updated row.",
        "operationId": "CatalogNodeVersions_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeVersion"
                }
              }
            }
          }
        },
        "summary": "Deprecate catalog node version",
        "tags": [
          "Catalog - node versions"
        ]
      }
    },
    "/catalog-node-sockets": {
      "post": {
        "operationId": "CatalogNodeSockets_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatalogNodeSocketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeSocket"
                }
              }
            }
          }
        },
        "summary": "Create input/output port on a catalog version",
        "tags": [
          "Catalog - node sockets"
        ]
      }
    },
    "/catalog-node-sockets/{id}": {
      "get": {
        "operationId": "CatalogNodeSockets_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeSocket"
                }
              }
            }
          }
        },
        "summary": "Get catalog node socket by id",
        "tags": [
          "Catalog - node sockets"
        ]
      },
      "patch": {
        "operationId": "CatalogNodeSockets_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatalogNodeSocketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeSocket"
                }
              }
            }
          }
        },
        "summary": "Update catalog node socket",
        "tags": [
          "Catalog - node sockets"
        ]
      },
      "delete": {
        "operationId": "CatalogNodeSockets_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Delete catalog node socket",
        "tags": [
          "Catalog - node sockets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Shake API",
    "description": "REST API for the **node catalog** (toolbox definitions) and **board graph** (runtime nodes, connections, props).\n\nEntity IDs are `bigint` in PostgreSQL and serialized as **strings** in JSON.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateCatalogNodeDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 128,
            "example": "tile"
          }
        },
        "required": [
          "slug"
        ]
      },
      "CatalogNode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "slug": {
            "type": "string",
            "example": "http-request",
            "maxLength": 128
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "slug"
        ]
      },
      "PaginatedMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20,
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "number",
            "example": 142,
            "minimum": 0
          },
          "totalPages": {
            "type": "number",
            "example": 8,
            "minimum": 0
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
      "PaginatedCatalogNodes": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogNode"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateCatalogNodeDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 128,
            "example": "tile"
          }
        }
      },
      "CatalogNodeVersion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "catalogNodeId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "version": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "HTTP Request v1",
            "maxLength": 255
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "deprecatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "catalogNodeId",
          "version",
          "name",
          "isActive",
          "deprecatedAt"
        ]
      },
      "PaginatedCatalogNodeVersions": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogNodeVersion"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CatalogNodeSocket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "type": {
            "type": "string",
            "enum": [
              "input",
              "output"
            ],
            "example": "input"
          },
          "name": {
            "type": "string",
            "example": "body",
            "maxLength": 255
          },
          "limit": {
            "type": "object",
            "example": 1,
            "nullable": true
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "catalogNodeVersionId",
          "type",
          "name"
        ]
      },
...
      "CreateCatalogNodeVersionDto": {
        "type": "object",
        "properties": {
          "catalogNodeId": {
            "type": "string",
            "example": "1",
            "description": "FK to catalog_nodes.id"
          },
          "version": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "example": "Tile"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "deprecatedAt": {
            "type": "string",
            "nullable": true,
            "example": "2026-01-01T00:00:00.000Z",
            "description": "When set, version is treated as deprecated in the toolbox"
          }
        },
        "required": [
          "catalogNodeId",
          "version",
          "name"
        ]
      },
      "UpdateCatalogNodeVersionDto": {
        "type": "object",
        "properties": {
          "catalogNodeId": {
            "type": "string",
            "example": "1",
            "description": "FK to catalog_nodes.id"
          },
          "version": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "example": "Tile"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "deprecatedAt": {
            "type": "string",
            "nullable": true,
            "example": "2026-01-01T00:00:00.000Z",
            "description": "When set, version is treated as deprecated in the toolbox"
          }
        }
      },
      "CreateCatalogNodeSocketDto": {
        "type": "object",
        "properties": {
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "FK to catalog_node_versions.id"
          },
          "type": {
            "type": "string",
            "enum": [
              "input",
              "output"
            ],
            "example": "input"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "example": "children"
          },
          "limit": {
            "type": "number",
            "nullable": true,
            "description": "Max connections allowed to this socket (null = unlimited)"
          }
        },
        "required": [
          "catalogNodeVersionId",
          "type",
          "name"
        ]
      },
      "UpdateCatalogNodeSocketDto": {
        "type": "object",
        "properties": {
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "FK to catalog_node_versions.id"
          },
          "type": {
            "type": "string",
            "enum": [
              "input",
              "output"
            ],
            "example": "input"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "example": "children"
          },
          "limit": {
            "type": "number",
            "nullable": true,
            "description": "Max connections allowed to this socket (null = unlimited)"
          }
        }
      },
...
    }
  }
}